### 2020-11-10 7970deb
updated documentation - license, contribution, tutorials
