+++ docs/source/conf.py
@@ -37,7 +37,9 @@ extensions = [
     	'sphinx.ext.intersphinx',
     	'sphinx.ext.ifconfig',
     	'sphinx.ext.viewcode',
-    	'sphinx.ext.githubpages'
+    	'sphinx.ext.githubpages',
+        'nbsphinx',
+        'sphinx.ext.mathjax',
 	]
 
 # Add any paths that contain templates here, relative to this directory.
