+++ pyinterpolate/data_processing/data_preparation/select_values_in_range.py
@@ -2,17 +2,18 @@ import numpy as np
 
 
 def select_values_in_range(data, lag, step_size):
-    """Function selects set of values which are greater than (lag - step size) and lesser than (lag + step size).
+    """
+    Function selects set of values which are greater than (lag - step size) and lesser than (lag + step size).
 
     INPUT:
 
     :param data: array of distances,
-    :param lag: (float) lag from which areas are included,
-    :param step_size: (float) step from lag to include in the analysis.
+    :param lag: (float) lag within areas are included,
+    :param step_size: (float) step between lags. Usually it is constant in each iteration and it is (0.5 * lag).
 
     OUTPUT:
 
-    :return: array mask with distances within specified radius.
+    :return: numpy array mask with distances within specified radius.
     """
 
     # Check if numpy array is given
