+++ docs/build/html/_sources/code_documentation/data preparation.ipynb.txt
@@ -1,0 +1,184 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Data preparation module\n",
+    "\n",
+    "Available methods:\n",
+    "\n",
+    "- get_points_within_area: prepares points data for further processing,\n",
+    "- prepare_areal_shapefile: prepares areal shapefile for processing and transforms it into numpy array,\n",
+    "- read_point_data: reads data and converts it into numpy array,\n",
+    "- select_values_in_range: selects set of values which are greater than (lag - step size) and lesser than (lag + step size),\n",
+    "- set_areal_weights: prepares array for weighted semivariance calculation."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## get_points_within_area\n",
+    "\n",
+    "```python\n",
+    "pyinterpolate.data_processing.data_preparation.get_points_within_area.get_points_within_area(area_shapefile, points_shapefile, areal_id_col_name, points_val_col_name, dropna=True, points_geometry_col_name='geometry', nans_to_zero=True)\n",
+    "```\n",
+    "\n",
+    "Function prepares points data for further processing.\n",
+    "\n",
+    "INPUT:\n",
+    "\n",
+    "- **area_shapefile**: (```string```) areal data ```shapefile``` address,\n",
+    "- **points_shapefile**: (```string```) points data ```shapefile``` address,\n",
+    "- **areal_id_col_name**: (```string```) name of the column with id of areas,\n",
+    "- **points_val_col_name**: (```string```) name of the value column of each point,\n",
+    "- **dropna***: (```bool```) if ```True``` then rows with ```NaN``` are deleted (areas without any points),\n",
+    "- **points_geometry_col_name***: (```string```) default is ```'geometry'``` as in ```GeoPandas GeoDataFrames```,\n",
+    "- **nans_to_zero**: (```bool```) if ```True``` then all ```NaN``` values are casted to 0.\n",
+    "\n",
+    "\n",
+    "OUTPUT:\n",
+    "\n",
+    "- ```numpy array``` of area id and array with point coordinates and values:\n",
+    "\n",
+    "```python\n",
+    "[\n",
+    "    area_id,\n",
+    "    [point_position_x, point_position_y, value]\n",
+    "]\n",
+    "```"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## prepare_areal_shapefile\n",
+    "\n",
+    "```python\n",
+    "pyinterpolate.data_processing.data_preparation.prepare_areal_shapefile.prepare_areal_shapefile(areal_file_address, id_column_name=None, value_column_name=None, geometry_column_name='geometry', dropnans=True)\n",
+    "```\n",
+    "\n",
+    "Function prepares areal shapefile for processing and transforms it into ```numpy array```. Function returns two lists.\n",
+    "\n",
+    "INPUT:\n",
+    "\n",
+    "- **areal_file_address**: (```string```) path to the shapefile with area data,\n",
+    "- **id_column_name**: (```string```) id column name, if not provided then index column is treated as the id,\n",
+    "- **value_column_name**: (```string```) value column name, if not provided then all values are set to ```NaN```,\n",
+    "- **geometry_column_name**: (```string```) default is ```'geometry'``` as in ```GeoPandas GeoDataFrames```,\n",
[…]
