+++ docs/build/html/_sources/code_documentation.rst.txt
@@ -1,0 +1,14 @@
+*************
+Documentation
+*************
+
+Welcome into the code documentation of the Pyinterpolate package. If you are looking for every method in the package it is the best place to go.
+
+Package modules:
+----------------
+
+.. toctree::
+   :maxdepth: 2
+
+   code_documentation/distances
+   code_documentation/data preparation
