+++ docs/build/html/_sources/index.rst.txt
@@ -13,6 +13,18 @@ Pyinterpolate
    introduction
    setup
    tutorials.rst
-   code-documentation.rst
+   code_documentation.rst
    contribution
    license
+
+===================================
+What can you do with Pyinterpolate?
+===================================
+
+PyInterpolate is designed as the Python library for geostatistics. It's role is to provide access to spatial statistics tools used in a wide range of studies. This package helps you **interpolate spatial data** with *Kriging* technique. In the close future you'll use more spatial interpolation tools.
+
+=====================
+Where should I start?
+=====================
+
+If you are first time here, we highly recommend to read introduction and learn about **Pyinterpolate** purpose. Otherwise it's a good idea to look into tutorials and quickly adapt code from there to your own projects.
