### 2020-11-19 1abe181
docs update data transform and viz
