+++ pyinterpolate/semivariance/areal_semivariance/areal_semivariance.py
@@ -13,18 +13,19 @@ from pyinterpolate.semivariance.areal_semivariance.block_to_block_semivariance.\
 
 
 class ArealSemivariance:
-    """Class for Areal Semivariance object which is used for the semivariogram deconvolution.
+    """
+    Class calculates semivariance of areas for Poisson Kriging (area to area and area to point).
 
     INITIALIZATION PARAMS:
 
-    param areal_data: (numpy array / list of lists) [area_id, area_geometry, centroid coordinate x,
-        centroid coordinate y, value],
-    :param areal_lags: (numpy array / list of lists) - array of lags (ranges of search),
+    :param areal_data: (numpy array / list) [area_id, area_geometry, centroid x,
+        centroid y, value],
+    :param areal_lags: (numpy array / list) - array of lags (ranges of search),
     :param areal_step_size: (float) step size for search radius,
     :param areal_points_data: (numpy array / list of lists)
         [area_id, [point_position_x, point_position_y, value]]
     :param weighted_semivariance: (bool) if False then each distance is treated equally when calculating
-        theoretical semivariance; if True then semivariances closer to the point have more weight,
+        theoretical semivariance; if True then semivariances closer to the point of origin have more weight,
     :param verbose: (bool) if True then all messages are printed, otherwise nothing.
     """
 
@@ -82,7 +83,8 @@ class ArealSemivariance:
 
     def regularize_semivariogram(self, within_block_semivariogram=None, between_blocks_semivariogram=None,
                                  empirical_semivariance=None, theoretical_semivariance_model=None):
-        """Function calculates regularized point support semivariogram in the form given in:
+        """
+        Function calculates regularized point support semivariogram in the form given in:
 
         Goovaerts P., Kriging and Semivariogram Deconvolution in the Presence of Irregular Geographical Units,
             Mathematical Geology 40(1), 101-128, 2008
@@ -100,12 +102,12 @@ class ArealSemivariance:
             y(va, va) and y(va+h, va+h) are the inblock semivariances of block a and block a+h separated
                 by the distance h weighted by the inblock population.
         :param between_blocks_semivariogram: semivariance between all blocks calculated from the theoretical model,
-        :param empirical_semivariance: (numpy array) empirical semivariance between areal centroids, default=None, if
+        :param empirical_semivariance: (numpy array) empirical semivariance between area centroids, default=None, if
             None is provided then empirical semivariance is computed by the _calculate_empirical_semivariance
-            method from areal centroids,
+            method from area centroids,
         :param theoretical_semivariance_model: (TheoreticalSemivariogram) theoretical semivariance model from
             TheoreticalSemivariance class, default is None, if None is provided then theoretical model is derived
-            from areal centroids and empirical semivariance.
+            from area centroids and empirical semivariance.
 
         OUTPUT:
 
@@ -269,7 +271,8 @@ class ArealSemivariance:
 
     def show_semivariograms(self):
         """
-        Function shows semivariograms calculated with the class.
+        Function shows semivariograms calculated by the class: Empirical semivariogram, Theoretical model,
+        Inblock Semivariance, Within-block semivariogram, Between blocks semivariogram, Regularized output.
         """
         plt.figure(figsize=(12, 12))
         plt.plot(self.experimental_semivariogram[:, 0], self.experimental_semivariogram[:, 1],
