+++ docs/build/html/_sources/code_documentation/data visualization.ipynb.txt
@@ -18,7 +18,10 @@
     "## interpolate_raster\n",
     "\n",
     "```python\n",
-    "pyinterpolate.data_visualization.interpolate_raster.interpolate_raster(data, dim=1000, number_of_neighbors=4, semivariogram_model=None)\n",
+    "pyinterpolate.data_visualization.interpolate_raster.interpolate_raster(data,\n",
+    "                                                                       dim=1000,\n",
+    "                                                                       number_of_neighbors=4,\n",
+    "                                                                       semivariogram_model=None)\n",
     "```\n",
     "\n",
     "Function interpolates raster from data points using ordinary kriging.\n",

[…]
