+++ docs/requirements.txt
@@ -1,0 +1,13 @@
+
+descartes==1.1.0
+geopandas==0.7.0
+matplotlib==3.2.1
+numpy==1.18.3
+tqdm==4.47.0
+pyproj==2.6.0
+scipy==1.4.1
+shapely==1.7.0
+fiona==1.8.13.post1; sys_platform=="darwin"
+rtree==0.9.4; sys_platform=="darwin"
+fiona==1.8; sys_platform=="linux"
+rtree>=0.8,<0.9; sys_platform=="linux"
