+++ docs/requirements.txt
@@ -11,3 +11,4 @@ fiona==1.8.13.post1; sys_platform=="darwin"
 rtree==0.9.4; sys_platform=="darwin"
 fiona==1.8; sys_platform=="linux"
 rtree>=0.8,<0.9; sys_platform=="linux"
+nbsphinx==0.8.0
