+++ docs/build/html/_sources/tutorials/Poisson Kriging - Area to Area (Advanced).ipynb.txt
@@ -58,10 +58,9 @@
     "import pandas as pd\n",
     "import geopandas as gpd\n",
     "\n",
-    "from pyinterpolate import prepare_areal_shapefile\n",
-    "from pyinterpolate import get_points_within_area\n",
-    "from pyinterpolate import TheoreticalSemivariogram\n",
-    "from pyinterpolate import ArealKriging"
+    "from pyinterpolate.io import prepare_areal_shapefile, get_points_within_area\n",
+    "from pyinterpolate.semivariance import TheoreticalSemivariogram\n",
+    "from pyinterpolate.kriging import ArealKriging"
    ]
   },
   {
