### 2020-11-26 #2869 @whedon
[REVIEW]: Pyinterpolate: Spatial Interpolation in Python for point measurements and aggregated datasets
**Submitting author:** @szymon-datalions (<a href="http://orcid.org/0000-0003-3525-2104">Szymon Moliński</a>)
**Repository:** <a href="https://github.com/szymon-datalions/pyinterpolate" target ="_blank">https://github.com/szymon-datalions/pyinterpolate</a>
**Version:** v0.2.0
