### 2020-12-15 cb22fcf
semivariance calculation only step size and max range params - code
