+++ pyinterpolate/semivariance/areal_semivariance/areal_semivariance.py
@@ -29,7 +29,7 @@ class ArealSemivariance:
     :param verbose: (bool) if True then all messages are printed, otherwise nothing.
     """
 
-    def __init__(self, areal_data, areal_lags, areal_step_size,
+    def __init__(self, areal_data, areal_step_size, max_areal_range,
                  areal_points_data, weighted_semivariance=False, verbose=False):
 
         """
@@ -37,8 +37,8 @@ class ArealSemivariance:
 
         :param areal_data: (numpy array / list of lists)
             [area_id, area_geometry, centroid coordinate x, centroid coordinate y, value],
-        :param areal_lags: (numpy array / list of lists) - array of lags (ranges of search),
         :param areal_step_size: (float) step size for search radius,
+        :param max_areal_range: (float) max distance to perform distance and semivariance calculations,
         :param areal_points_data: (numpy array / list of lists)
             [area_id, [point_position_x, point_position_y, value]]
         :param weighted_semivariance: (bool) if False then each distance is treated equally when calculating
@@ -49,8 +49,9 @@ class ArealSemivariance:
         # Initial data
         self.areal_dataset = areal_data
         self.areal_centroids = self.areal_dataset[:, 2:]
-        self.areal_lags = areal_lags
+        self.areal_lags = np.arange(0, max_areal_range, areal_step_size)
         self.areal_ss = areal_step_size
+        self.areal_max_range = max_areal_range
         self.within_area_points = areal_points_data
         self.weighted_semivariance = weighted_semivariance
 
@@ -69,7 +70,7 @@ class ArealSemivariance:
     def _calculate_empirical_semivariance(self):
         """Method calculates theoretical semivariance between areal centroids
         :return gamma: (float) semivariance between areal centroids."""
-        gamma = calculate_semivariance(self.areal_centroids, self.areal_lags, self.areal_ss)
+        gamma = calculate_semivariance(self.areal_centroids,  self.areal_ss, self.areal_max_range)
         return gamma
 
     def _calculate_theoretical_semivariance(self):
