### 2020-12-23 a470ea5
simplified ranges calculations - tests and notebooks (docs still missing); changed io module to io_ops
