+++ pyinterpolate/test/semivariance/test_calculate_semivariance_within_blocks.py
@@ -5,8 +5,8 @@ from pyinterpolate.semivariance.areal_semivariance.within_block_semivariance.cal
 
 import numpy as np
 import geopandas as gpd
-from pyinterpolate.io.prepare_areal_shapefile import prepare_areal_shapefile
-from pyinterpolate.io.get_points_within_area import get_points_within_area
+from pyinterpolate.io_ops.prepare_areal_shapefile import prepare_areal_shapefile
+from pyinterpolate.io_ops.get_points_within_area import get_points_within_area
 from pyinterpolate.semivariance.semivariogram_estimation.calculate_semivariance import calculate_semivariance
 from pyinterpolate.semivariance.semivariogram_fit.fit_semivariance import TheoreticalSemivariogram
 
@@ -56,7 +56,7 @@ class TestCalculateSemivarianceWithinBlocks(unittest.TestCase):
         inblock_semivariance = np.array(inblock_semivariance)
 
         data_point = inblock_semivariance[inblock_semivariance[:, 0] == 1][0]
-        self.assertEqual(int(data_point[1]), 111, "First data point's integer part should be equal to 111")
+        self.assertEqual(int(data_point[1]), 104, "First data point's integer part should be equal to 104")
 
 
 if __name__ == '__main__':
