+++ pyinterpolate/test/semivariance/test_fit_semivariance.py
@@ -1,7 +1,7 @@
 import unittest
 import os
 import numpy as np
-from pyinterpolate.io.read_data import read_point_data
+from pyinterpolate.io_ops.read_data import read_point_data
 from pyinterpolate.distance.calculate_distances import calc_point_to_point_distance
 from pyinterpolate.semivariance.semivariogram_fit.fit_semivariance import TheoreticalSemivariogram
 from pyinterpolate.semivariance.semivariogram_estimation.calculate_semivariance import calculate_semivariance
@@ -37,11 +37,11 @@ class TestFitSemivariance(unittest.TestCase):
         t_non_weighted = TheoreticalSemivariogram(dataset, gamma_non)
         t_weighted = TheoreticalSemivariogram(dataset_weights[:, :-1], gamma_w)
 
-        model_non_weighted = t_non_weighted.find_optimal_model(weighted=False, number_of_ranges=20)
-        model_weighted = t_weighted.find_optimal_model(weighted=False, number_of_ranges=20)
+        model_non_weighted = t_non_weighted.find_optimal_model(weighted=False, number_of_ranges=20)  # linear
+        model_weighted = t_weighted.find_optimal_model(weighted=False, number_of_ranges=20)  # linear
 
-        self.assertEqual(model_non_weighted, 'exponential', "Non-weighted model should be exponential")
-        self.assertEqual(model_weighted, 'exponential', "Weighted model should be exponential")
+        self.assertEqual(model_non_weighted, 'linear', "Non-weighted model should be linear")
+        self.assertEqual(model_weighted, 'linear', "Weighted model should be linear")
 
     def test_fit_semivariance_io(self):
         # Prepare fake model for fit semivariance class
