### 2020-12-25 #2869 @kenohori
Hello!

I started looking at your submission a few days ago and decided that some initial feedback would be useful, especially regarding the paper. In short, I like the summary and statement of need, which are fine and clear. However, the rest of the paper is a general introduction to kriging, which doesn't say much about how it's performed in your software. I would recommend adding a short description of the different types of kriging you do and how they're implemented.

Also, your paper is missing references to the state of the field. What are the main (open or closed) alternatives out there? How does your software compare to them?

Regarding the software itself, I haven't finished testing it yet, but everything I've tested seems to work (with some very minor issues). However, in the meantime I've noticed a few things.

- Installation/setup: it would be nice to separate the dependencies are handled by pip and those that need to be installed separately.
- The functions documentation in the readme actually contains many things, including tutorials on how to use the software
- How can I test that the software is working correctly? Do you have some unit tests or reference output?
