### 2020-12-26 9d2ffa9
lags parameter removed; documentation updated
