+++ docs/build/html/_sources/code_documentation/transform.ipynb.txt
@@ -314,15 +314,13 @@
     "    step_size)\n",
     "```\n",
     "\n",
-    "Function selects set of values which are greater than (```lag - step size```) and lesser than (```lag + step size```).\n",
+    "Function selects set of values which are greater than (```lag - step size / 2```) and smaller or equal than (```lag + step size / 2```).\n",
     "\n",
     "INPUT:\n",
     "\n",
     "- **data**: (_numpy array_) distances,\n",
     "- **lag**: (_float_) lag within areas are included,\n",
-    "- **step_size**: (_float_) step between lags. Usually it is constant in each iteration and it is:\n",
-    "\n",
-    "$$0.5 * lag$$\n",
+    "- **step_size**: (_float_) step between lags.\n",
     "\n",
     "\n",
     "OUTPUT:\n",
@@ -388,7 +386,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.7.8"
+   "version": "3.7.3"
   }
  },
  "nbformat": 4,
