### 2020-12-27 #2869 @szymon-datalions
Hi @kenohori!

Thank you for your feedback! Thanks for suggestions too, I'll describe implementation of Kriging techniques and I'll compare package to other programs available on Github or elsewhere. For the rest of your recommendations:

> Installation/setup: it would be nice to separate the dependencies are handled by pip and those that need to be installed separately.

I don't understand what do you mean by that? Should I emphasize those installation requirements in two separate files, re-arrange actual documentation or maybe should I do something else (with installation itself?)

> How can I test that the software is working correctly? Do you have some unit tests or reference output?

In the main module there is a test module for checking how functions behave (unit tests) => `pyinterpolate/test`. Usually I'm checking every change with available tutorials too, but it's more manual work instead of automated tests.
