### 2020-12-30 #2869 @kenohori
> I don't understand what do you mean by that? Should I emphasize those installation requirements in two separate files, re-arrange actual documentation or maybe should I do something else (with installation itself?)

My suggestion would be to summarise the instructions from the manual setup file in the main readme. It could be something like:

1. install libspatialindex (eg with apt-get on Linux or brew on Mac)
2. `pip install pyinterpolate`
3. run tests

> In the main module there is a test module for checking how functions behave (unit tests) => pyinterpolate/test. Usually I'm checking every change with available tutorials too, but it's more manual work instead of automated tests.

Thanks! I'll have a look at those. It would be nice to add this information to the Readme and paper as well.
