### 2021-01-01 1bba77c
check if observations with the same coordinates are in a dataset
