### 2021-01-01 fb7bd71
corrected unique coordinate checking function
