### 2021-01-08 880b65e
variogram cloud calculations - initial function
