+++ pyinterpolate/semivariance/semivariogram_estimation/calculate_semivariance.py
@@ -1,7 +1,67 @@
+from collections import OrderedDict
+
 import numpy as np
 
 from pyinterpolate.distance.calculate_distances import calc_point_to_point_distance
 from pyinterpolate.transform.select_values_in_range import select_values_in_range
+
+import matplotlib.pyplot as plt
+
+
+def build_variogram_point_cloud(data, step_size, max_range):
+    """
+    Function calculates variogram point cloud of a given set of points for
+        a given set of distances. Variogram is calculated as a squared difference
+        of each point against other point within range specified by step_size
+        parameter. 
+
+    INPUT:
+
+    :param data: (numpy array) coordinates and their values,
+    :param step_size: (float) step size of circle within radius are analyzed points,
+    :param max_range: (float) maximum range of analysis.
+
+    OUTPUT:
+
+    :return: variogram_cloud - dict with pairs {lag: list of squared differences}.
+    """
+
+    distances = calc_point_to_point_distance(data[:, :-1])
+    lags = np.arange(0, max_range, step_size)
+    variogram_cloud = OrderedDict()
+
+    # Calculate squared differences
+    # They are halved to be compatibile with semivariogram
+
+    for h in lags:
+        distances_in_range = select_values_in_range(distances, h, step_size)
+        sem = (data[distances_in_range[0], 2] - data[distances_in_range[1], 2]) ** 2
+        sem = sem / 2
+        variogram_cloud[h] = sem
+
+    return variogram_cloud
+
+
+def show_variogram_cloud(variogram_cloud, figsize=None):
+    """
+    Function shows boxplots of variogram lags. It is especially useful when
+        you want to check outliers in your dataset.
+
+    INPUT:
+
+    :param variogram_cloud: (OrderedDict) lags and halved squared differences between
+       points,
+    :param figsize: (tuple) figure size (width, height).
+    """
+    if figsize is None:
+        figsie = (14, 6)
+
+    data = [x for x in variogram_cloud.values()]
+    fig, ax = plt.subplots(figsize=figsize)
+    ax.boxplot(data)
+    xtickNames = plt.setp(ax, xticklabels=variogram_cloud.keys())
+    plt.setp(xtickNames, rotation=45, fontsize=8)
+    plt.show()
 
 
 def calculate_semivariance(data, step_size, max_range):
