### 2021-01-16 a2fdbb2
semivariogram point cloud - tests and baseline notebook
