+++ pyinterpolate/semivariance/__init__.py
@@ -2,4 +2,5 @@ from .areal_semivariance.areal_semivariance import ArealSemivariance
 from .semivariogram_deconvolution.regularize_semivariogram import RegularizedSemivariogram
 from .semivariogram_estimation.calculate_covariance import calculate_covariance
 from .semivariogram_estimation.calculate_semivariance import calculate_semivariance, calculate_weighted_semivariance
+from .semivariogram_estimation.calculate_semivariance import build_variogram_point_cloud, show_variogram_cloud
 from .semivariogram_fit.fit_semivariance import TheoreticalSemivariogram
