### 2021-01-17 4ded7b6
variogram point cloud tests, tutorial and docs
