+++ pyinterpolate/semivariance/__init__.py
@@ -3,4 +3,6 @@ from .semivariogram_deconvolution.regularize_semivariogram import RegularizedSem
 from .semivariogram_estimation.calculate_covariance import calculate_covariance
 from .semivariogram_estimation.calculate_semivariance import calculate_semivariance, calculate_weighted_semivariance
 from .semivariogram_estimation.calculate_semivariance import build_variogram_point_cloud, show_variogram_cloud
+from .semivariogram_estimation.calculate_semivariance import calc_semivariance_from_pt_cloud
+from .semivariogram_estimation.calculate_semivariance import remove_outliers
 from .semivariogram_fit.fit_semivariance import TheoreticalSemivariogram
