### 2021-01-18 #2869 @chrisbrunsdon
Apologies for the delays here - I'm currently involved in working on covid data analysis for the Irish govt - and things have recently got busy.  Will start responding soon...   Just to report that I have successfully downloaded and run the module without any major problems,  and that suggestions mainly relate to examples given...
