+++ paper/paper.bib
@@ -69,3 +69,67 @@ doi="10.1007/978-3-319-78999-6_29",
 url="https://doi.org/10.1007/978-3-319-78999-6_29"
 }
 
+@software{benjamin_murphy_2020_3991907,
+  author       = {Benjamin Murphy and
+                  Sebastian Müller and
+                  Roman Yurchak},
+  title        = {GeoStat-Framework/PyKrige v1.5.1},
+  month        = aug,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v1.5.1},
+  doi          = {10.5281/zenodo.3991907},
+  url          = {https://doi.org/10.5281/zenodo.3991907}
+}
+
+@article{scikit-learn,
+ title={Scikit-learn: Machine Learning in {P}ython},
+ author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ journal={Journal of Machine Learning Research},
+ volume={12},
+ pages={2825--2830},
+ year={2011}
+}
+
+@Manual{GRASS_GIS_software,
+    title = {Geographic Resources Analysis Support System (GRASS GIS) Software},
+    author = {{GRASS Development Team}},
+    organization = {Open Source Geospatial Foundation},
+    address = {USA},
+    year = {2020},
+    url = {https://grass.osgeo.org}
+}
+
+@software{eli_knaap_2020_4385980,
+  author       = {eli knaap and
+                  Renan Xavier Cortes and
+                  Sergio Rey and
+                  James Gaboardi and
+                  Patty Frontiera},
+  title        = {pysal/tobler: Release v0.5.4},
+  month        = dec,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v0.5.4},
+  doi          = {10.5281/zenodo.4385980},
+  url          = {https://doi.org/10.5281/zenodo.4385980}
+}
+
+@article{PEBESMA2004683,
+title = "Multivariable geostatistics in S: the gstat package",
+journal = "Computers & Geosciences",
+volume = "30",
+number = "7",
+pages = "683 - 691",
+year = "2004",
+issn = "0098-3004",
+doi = "https://doi.org/10.1016/j.cageo.2004.03.012",
+url = "http://www.sciencedirect.com/science/article/pii/S0098300404000676",
+author = "Edzer J Pebesma",
+keywords = "Kriging, Cokriging, Linear model of coregionalisation, Open source software, S language, Stochastic simulation",
[…]
