### 2021-01-28 #2869 @kenohori
Hello!

Just finished testing the software. As far as I can tell, every function in the software works. All unit tests are okay and the functionality in the tutorials seems to work fine.

However, there are some minor issues in the tutorials in v0.2.0, which I guess are due to name changes in code refactoring.
* In semivariogram regularization: calculate_semivariance.calculate_semivariance -> calculate_semivariance
* Poisson area to point: s_radius -> max_search_radius

Looking forward to the updated paper!
