+++ paper/paper.md
@@ -79,33 +79,21 @@ Pyinterpolate is designed from seven modules and they cover all operations neede
 
 ![Structure of Pyinterpolate package.\label{fig1}](fig1.png)
 
-Modules follow data processing and modeling steps. The first module is **io_ops** which reads point data from text files and areal or point data from shapefiles, then changes data structure for further processing. **Transform** module is responsible for all tasks related to changes in data structure during program execution. Sample tasks are:
+Modules follow typical data processing and modeling steps. The first module is **io_ops** which reads point data from text files and areal or point data from shapefiles, then changes data structure for further processing. **Transform** module is responsible for all tasks related to changes in data structure during program execution. Sample tasks are:
 
 - finding centroids of areal data,
 - building masks of points within lag.
 
-Functions for distance calculation between points and between areas (blocks) are grouped within **distance** module. **Semivariance** is most complex part of Pyinterpolate package. It has three special classes:
-
--	```TheoreticalSemivariogram```: class for theoretical semivariance model calculations and storage,
--	```ArealSemivariance```: class for areal semivariance models calculations and storage,
--	```RegularizedSemivariogram```: class which performs semivariogram regularization and stores regularized semivariogram.
-
-**Semivariance** module has other functions important for spatial analysis: 
+Functions for distance calculation between points and between areas (blocks) are grouped within **distance** module. **Semivariance** is most complex part of Pyinterpolate package. It has three special classes for calculation and storage of different types of semivariograms (experimental, theoretical, areal and point types). **Semivariance** module has other functions important for spatial analysis: 
 
 - function for experimental semivariance / covariance calculation, 
 - weighted semivariance estimation,
 - variogram cloud preparation, 
 - outliers removal.
 
-**Kriging** module contains three main types of models:
+**Kriging** module contains three main types of models Ordinary and Simple Kriging models as well Poisson Kriging of areal counts models. Areal models are derived from [@Goovaerts:2008], simple Kriging and ordinary Kriging models are based on [@Armstrong:1998].
 
--	```ArealKriging```: for *area-to-area* and *area-to-point* Poisson Kriging,
--	```CentroidPoissonKriging```: for Poisson Kriging of areal centroids,
--	```Krige```: for ordinary and simple kriging of points.
-
-```ArealKriging``` and ```CentroidPoissonKriging``` models are derived from [@biblio], simple Kriging and ordinary Kriging models are based on [@ biblio].
-
-Package has two special functions in **viz** and **misc** module. **Viz** module has ```interpolate_raster()``` function for creation of rasterized numpy array of evenly spaced points (visualization purpose). **Misc** module offers special function for automatic comparison of multiple Kriging models on the same dataset (area-to-area Poisson Kriging, Centroid-based Poisson Kriging, Ordinary Kriging, Simple Kriging). Evaluation metric of this method is an average root mean squared error over multiple random divisions of passed dataset.
+It is possible to show output as numpy array with **viz** module and to compare multiple kriging models on the same dataset with **misc** module. Evaluation metric for comparison is an average root mean squared error over multiple random divisions of a passed dataset.
 
 # Example use case: Breast Cancer Rate in Pennsylvania State
 
@@ -118,112 +106,17 @@ Package logic follows typical pipeline:
 
[…]
