+++ paper/paper.bib
@@ -133,3 +133,34 @@ keywords = "Kriging, Cokriging, Linear model of coregionalisation, Open source s
 abstract = "This paper discusses advantages and shortcomings of the S environment for multivariable geostatistics, in particular when extended with the gstat package, an extension package for the S environments (R, S-Plus). The gstat S package provides multivariable geostatistical modelling, prediction and simulation, as well as several visualisation functions. In particular, it makes the calculation, simultaneous fitting, and visualisation of a large number of direct and cross (residual) variograms very easy. Gstat was started 10 years ago and was released under the GPL in 1996; gstat.org was started in 1998. Gstat was not initially written for teaching purposes, but for research purposes, emphasising flexibility, scalability and portability. It can deal with a large number of practical issues in geostatistics, including change of support (block kriging), simple/ordinary/universal (co)kriging, fast local neighbourhood selection, flexible trend modelling, variables with different sampling configurations, and efficient simulation of large spatially correlated random fields, indicator kriging and simulation, and (directional) variogram and cross variogram modelling. The formula/models interface of the S language is used to define multivariable geostatistical models. This paper introduces the gstat S package, and discusses a number of design and implementation issues. It also draws attention to a number of papers on integration of spatial statistics software, GIS and the S environment that were presented on the spatial statistics workshop and sessions during the conference Distributed Statistical Computing 2003."
 }
 
+@online{cancerData,
+  author = {National Cancer Institute},
+  title = {Incidence Rates Table: Breast Cancer: Pennsylvania State},
+  year = 2021,
+  url = {https://www.statecancerprofiles.cancer.gov/incidencerates/index.php?stateFIPS=42&areatype=county&cancer=055&race=00&sex=2&age=001&stage=999&year=0&type=incd&sortVariableName=rate&sortOrder=default&output=0#results},
+  urldate = {2021-02-03}
+}
+
+@online{popCensus,
+  author = {United States Census Bureau},
+  title = {Centers of Population for the 2010 Census},
+  year = 2021,
+  url = {https://www.census.gov/geographies/reference-files/2010/geo/2010-centers-population.html},
+  urldate = {2021-02-03}
+}
+
+@online{pennSpatial,
+  author = {Pennsylvania Spatial Data Access (PASDA)},
+  title = {Pennsylvania county boundaries},
+  year = 2021,
+  url = {https://www.pasda.psu.edu/uci/DataSummary.aspx?dataset=24},
+  urldate = {2021-02-03}
+}
+
+@online{paperRepo,
+  author = {S. Moliński},
+  title = {Pyinterpolate: paper materials},
+  year = 2021,
+  url = {https://github.com/szymon-datalions/pyinterpolate-paper},
+  urldate = {2021-02-03}
+}
