### 2021-02-03 #2869 @szymon-datalions
Hi @kenohori , Hi @chrisbrunsdon ,

I've updated paper and README.md in the ```paper``` branch of package repository. I'll check all namespaces in tutorials but I think that in current version (0.2.2) every tutorial should pass. Moreover, I'm changing areal dataset for tutorials because current dataset (road accidents) is not spatially correlated. It will be done in a day or two.

Thank you a lot for your feedback!
