### 2021-03-02 #2869 @szymon-datalions
Hi,

I hope that you're doing well in this COVID time. I'm writing to remind you about the paper. I'm going to update package during this week, along with it tutorials (examples) and  documentation will be updated. I've changed example with _road accidents_ data (not spatially correlated at the scale of interest) to _Breast Cancer Cases in Pennsylvania State_ counties where spatial correlation occurs. Here are the most recent tutorials: https://github.com/szymon-datalions/pyinterpolate/tree/dev/tutorials

Thanks!
