+++ MANIFEST.in
@@ -1,0 +1,2 @@
+exclude command.sh
+exclude readthedocs.yml
