### 2021-03-10 d2f796b
dir semivar calculations - scripts
