### 2021-03-10 da8785a
directional semivariograms methods
