+++ pyinterpolate/test/semivariance/test_calculate_semivariance.py
@@ -19,9 +19,45 @@ class TestCalculateDirectionalSemivariance(unittest.TestCase):
     def _test_case_1(self, data, step, rng, angle):
         semivariance = calculate_directional_semivariogram(data, step, rng, angle)
         equals = semivariance[1:, -1] == 13
-        self.assertTrue(equals.all())
-        self.assertCountEqual(semivariance[0], [0, 0, 0])
+        self.assertTrue(equals.all(), 'Number of points is not equal, should be 13 for each lag but 0')
+        self.assertCountEqual(semivariance[0], [0, 0, 0], 'Zero lag should have 0 element and 0 semivariance')
 
+    def _test_case_2(self, data, step, rng, angle):
+        semivariance = calculate_directional_semivariogram(data, step, rng, angle)
+        output = [[4.625, 24],
+                  [6, 20],
+                  [3.125, 16]]
+        t1 = semivariance[2, 1:] == output[0]
+        t2 = semivariance[4, 1:] == output[1]
+        t3 = semivariance[6, 1:] == output[2]
+        t = np.array([t1, t2, t3]).all()
+        self.assertTrue(t, 'Semivariance values are not equal expected results')
+
+    def _test_case_3(self, data, step, rng, angle1, angle2):
+        semivariance_ew = calculate_directional_semivariogram(data, step, rng, angle1)
+        semivariance_we = calculate_directional_semivariogram(data, step, rng, angle2)
+
+        t = (semivariance_ew == semivariance_we).all()
+        self.assertTrue(t, 'Semivariance should be the same for direction and its opposite direction')
+
+    def _test_case_4(self, data, step, rng, angle, tolerance):
+        semivariance = calculate_directional_semivariogram(data, step, rng, angle, tolerance)
+        output = [1.1, 7.31, 100]
+        lags = [0, 100, 296, 340, 542, 170, 564, 146, 488, 252]
+
+        t1 = (semivariance[1] == output).all()
+        self.assertTrue(t1, 'Semivariance output for the first lag should be equal to: [1.1, 7.31, 100]')
+
+        t2 = (semivariance[:, -1] == lags).all()
+        self.assertTrue(t2, 'Number of pairs per lag is not equal expected number of pairs')
+
+    def _test_case_5(self, data, step, rng, angle):
+        self.assertRaises(ValueError, calculate_directional_semivariogram,
+                          data, step, rng, angle)
+
+    def _test_case_6(self, data, step, rng, angle, tol):
+        self.assertRaises(ValueError, calculate_directional_semivariogram,
+                          data, step, rng, angle, tol)
 
     def test_calculate_directional_semivariogram(self):
 
@@ -39,17 +75,28 @@ class TestCalculateDirectionalSemivariance(unittest.TestCase):
         ew_dir = 90
         we_dir = 270
         ne_sw_dir = 30
-        nw_se_dir = 200
-        wrong_dir_too_much = 400
+        tol = 0.5
+        tol_wrong = 1.1
         wrong_dir_too_low = -1
 
         # Test simple arr
         # NS direction tests: semivariance constant across all lags
         self._test_case_1(simple_arr, step_size, max_range, ns_dir)
 
-        # EW 
+        # EW direction tests
+        self._test_case_2(simple_arr, step_size, max_range, ew_dir)
 
+        # EW == WE test case
+        self._test_case_3(simple_arr, step_size, max_range, ew_dir, we_dir)
 
+        # NE-SW test case
+        self._test_case_4(complex_arr, step_size, max_range, ne_sw_dir, tol)
+
+        # Wrong dir case
+        self._test_case_5(complex_arr, step_size, max_range, wrong_dir_too_low)
+
[…]
