+++ README.md
@@ -34,7 +34,7 @@ Package allows o perform multiple spatial interpolation tasks. The flow of analy
 **[1.]** Read and prepare data.
 
 ```python
-from pyinterpolate.io import read_point_data
+from pyinterpolate.io_ops import read_point_data
 
 point_data = read_point_data('xyz_txt_file.txt')
 ```

[…]
