### 2021-03-10 #2869 @szymon-datalions
Hi,

I forgot to update paper here, but many changes were made! Additionally, new version of package was updated to PyPi.
