### 2021-03-13 #2869 @kenohori
Hello @szymon-datalions,

I've had a look at the revised paper. First of all, thank you for adding the sections on the state of the art, the types of kriging performed in your software and the new use case. I think the paper is much more complete and easier to follow.

That being said, I'm honestly a bit unsure about what to make of the new breast cancer use case. I'm (very much) not a medical doctor, but is there a reason to expect spatial correlation here? Maybe something environmental? Intuitively, I'd think that spatial correlation would be pretty weak, and the semivariograms indeed reach the sill very quickly. To me, it's maybe not the ideal use case to showcase your package.

Speaking about those semivariograms, I noticed that the distances appear to be in degrees. This means that they're not actual distances on the ground, so I think the dataset needs to be projected. Something like NAD 83 should be fine.

Finally, I think it would be very helpful to have a single section (with a diagram?) that clearly describes the types of kriging you can do, their requirements and the steps to perform them. Currently, that information is there, but it's dispersed throughout the text (in the statement of need, the spatial interpolation section and the structure of package).
