### 2021-03-13 #2869 @szymon-datalions
Hi @kenohori,

I'm extremely grateful for your comments!

You have right that spatial correlation in the case of cancer is weak. Usually Poisson Kriging in the case of public health studies serves rather as a smoother (outliers removal) and/or a kind of filter to avoid visual bias of choropleth maps related to the different population density patterns within administrative units. Anyway, you made mi think that example is not clear so I'm going with your suggestion and I'll change data to depict interpolation instead of filtering problem, maybe I'll show two examples instead of one and I'll pick different data source for it.

What to do with distances is clear. Diagram section is a great idea and I think that paper will be much more readable. Maybe I'll change few things in its structure to avoid unnecessary chaos.
