### 2021-03-17 6f9e6f2
import semivariograms method #1
