### 2021-03-17 409e11e
new method for Theoretical Semivariogram
