+++ pyinterpolate/test/semivariance/test_fit_semivariance.py
@@ -1,7 +1,8 @@
 import unittest
 import os
 import numpy as np
-from pyinterpolate.distance.calculate_distances import calc_point_to_point_distance
+import pandas as pd
+
 from pyinterpolate.semivariance.semivariogram_fit.fit_semivariance import TheoreticalSemivariogram
 from pyinterpolate.semivariance.semivariogram_estimation.calculate_semivariance import calculate_semivariance
 from pyinterpolate.semivariance.semivariogram_estimation.calculate_semivariance import calculate_weighted_semivariance
@@ -15,7 +16,7 @@ class TestFitSemivariance(unittest.TestCase):
         my_dir = os.path.dirname(__file__)
         path = os.path.join(my_dir, '../sample_data/armstrong_data.npy')
 
-        self.complex_arr = np.load(path)
+        self.dataset = np.load(path)
         self.step_size = 1.1
         self.max_range = 10
 
@@ -27,26 +28,19 @@ class TestFitSemivariance(unittest.TestCase):
         dataset_weights[:, :-1] = self.dataset
         dataset_weights[:, -1] = new_col
 
-        # Set semivariance params
-        distances = calc_point_to_point_distance(dataset_weights[:, :-2])
-
-        maximum_range = np.max(distances)
-        number_of_divisions = 10
-        step_size = maximum_range / number_of_divisions
-
         # Calculate weighted and non-weighted semivariance
 
-        gamma_w = calculate_weighted_semivariance(dataset_weights, step_size, maximum_range)
-        gamma_non = calculate_semivariance(self.dataset, step_size, maximum_range)
+        gamma_w = calculate_weighted_semivariance(dataset_weights, self.step_size, self.max_range)
+        gamma_non = calculate_semivariance(self.dataset, self.step_size, self.max_range)
 
         # Fit semivariance - find optimal models
         t_non_weighted = TheoreticalSemivariogram(self.dataset, gamma_non)
         t_weighted = TheoreticalSemivariogram(dataset_weights[:, :-1], gamma_w)
 
-        model_non_weighted = t_non_weighted.find_optimal_model(weighted=False, number_of_ranges=20)  # linear
-        model_weighted = t_weighted.find_optimal_model(weighted=False, number_of_ranges=20)  # linear
+        model_non_weighted = t_non_weighted.find_optimal_model(weighted=False, number_of_ranges=8)  # linear
+        model_weighted = t_weighted.find_optimal_model(weighted=False, number_of_ranges=8)  # linear
 
-        self.assertEqual(model_non_weighted, 'linear', "Non-weighted model should be linear")
+        self.assertEqual(model_non_weighted, 'exponential', "Non-weighted model should be exponential")
         self.assertEqual(model_weighted, 'linear', "Weighted model should be linear")
 
     def test_fit_semivariance_io(self):
@@ -81,7 +75,22 @@ class TestFitSemivariance(unittest.TestCase):
         self.assertEqual(fake_theoretical_smv.chosen_model_name, fmn, "Model name should be linear")
 
     def test_semivariance_export(self):
-        pass
+        gamma = calculate_semivariance(self.dataset, self.step_size, self.max_range)
+        theo_model = TheoreticalSemivariogram(self.dataset, gamma)
+        theo_model.find_optimal_model(number_of_ranges=8)
+        filepath = '../sample_data/test_semivariance_export.csv'
+        theo_model.export_semivariance(filepath)
+        df = pd.read_csv(filepath)
+
+        columns = ['lag', 'experimental', 'theoretical']
+        for c in columns:
+            self.assertIn(c, df.columns, f'DataFrame is corrupted, missing {c} column')
+
[…]
