### 2021-03-17 4384e89
new method for Theoretical Semivariogram - bug with import corrected
