+++ paper/paper.md
@@ -4,8 +4,9 @@ tags:
   - Python
   - spatial statistics
   - spatial interpolation
-  - Kriging
+  - Poisson Kriging
   - Semivariogram Deconvolution
+  - public health
 authors:
   - name: Szymon Moliński
     orcid: 0000-0003-3525-2104
@@ -19,15 +20,171 @@ bibliography: paper.bib
 
 # Summary
 
-Spatial Interpolation techniques are used to retrieve data at unknown locations from point observations and/or areal aggregates. The basic idea is that measurement at one location is a function of weighted distance between this place and its neighbors. It is known as the Tobler's First Law of Geography, which states: *everything is related to everything else, but near things are more related than distant things* [@Tobler:1970]. Important method of spatial interpolation is Kriging, which initially was designed for mining applications. Today this technique is much more complex, and it can be applied to problems of environmental science, hydrogeology, natural resources monitoring, remote sensing and epidemiology [@Chilès:2018]. Kriging is usually applied to point measurements and regular blocks, but nowdays it could be used with areal data of irregular shapes and sizes [@Goovaerts:2007].
+Spatial Interpolation techniques are used to interpolate values at unknown locations and/or filter and smooth existing data sources. Those methods work for point observations and areal aggregates. The basic idea is that measurement at one location is a function of weighted distance between this place and its neighbors. It is known as the Tobler's First Law of Geography, which states: *everything is related to everything else, but near things are more related than distant things* [@Tobler:1970]. Kriging technique designed for mining applications exploits this statement formally and nowadays it has gained a lot of attention outside the initial area of interest. Kriging is rather a set of methods which can be applied to environmental science, hydrogeology, natural resources monitoring, remote sensing, epidemiology and ecology problems [@Chilès:2018]. Kriging is usually used with point measurements or regular block units but many real-world datasets are different. Especially challenging are measurements of rates over areas of irregular shapes and sizes, as example administrative units in every country [@Goovaerts:2007]. Pyinterpolate is designed to tackle this problem and to simplify area variogram regularization and analysis for social, environmental, public health and similar sciences. Package merges basic Kriging techniques with more sophisticated Poisson Kriging methods.
 
 # Statement of need
 
-Pyinterpolate is a Python package for spatial interpolation and it is designed to perform predictions from point measurements and areal aggregates of different sizes and shapes. Pyinterpolate automates tasks performed by spatial statisticians, it helps with data exploration, semivariogram estimation and kriging predictions. It allows researchers to perform semivariogram deconvolution in presence of irregular administrative units and to apply Poisson Kriging to areal counts. This type of Kriging is widely used in the social sciences, epidemiology and spatial statistics [@Goovaerts:2007; @Goovaerts:2008; @Kerry:2013].
+Pyinterpolate is a Python package for spatial interpolation and it is designed to perform predictions from point measurements and areal aggregates of different sizes and shapes. Pyinterpolate automates tasks performed by spatial statisticians, it helps with data exploration, semivariogram estimation and kriging predictions. Thing which makes Pyinterpolate different from other spatial interpolation packages is an ability to perform Kriging of areas of different shapes and sizes which is extremely important in the context of social, medical and ecological sciences.
 
[…]
