+++ paper/paper.bib
@@ -5,6 +5,17 @@
   year = {1998}
 }
 
+@article{Goovaerts:2006,
+  author = {Goovaerts, Pierre},
+  month = {11},
+  title = {Geostatistical analysis of disease data: accounting for spatial support and population density in the isopleth mapping of cancer mortality risk using area-to-point Poisson kriging},
+  doi = {10.1186/1476-072X-5-52},
+  urldate = {2021-03-24},
+  volume = {5},
+  year = {2006},
+  journal = {International Journal of Health Geographics}
+}
+
 @article{Goovaerts:2007,
   author = {Goovaerts, Pierre},
   month = {12},
