### 2021-03-27 d2c347a
inline equation corrections
