+++ paper/paper.bib
@@ -167,11 +167,3 @@ abstract = "This paper discusses advantages and shortcomings of the S environmen
   url = {https://www.pasda.psu.edu/uci/DataSummary.aspx?dataset=24},
   urldate = {2021-02-03}
 }
-
-@online{paperRepo,
-  author = {S. Moliński},
-  title = {Pyinterpolate: paper materials},
-  year = 2021,
-  url = {https://github.com/szymon-datalions/pyinterpolate-paper},
-  urldate = {2021-02-03}
-}
