+++ paper/paper.md
@@ -95,11 +95,11 @@ Table 1:
 
 | Kriging Type | Input | Output | Steps | Comment |
 |------------------|---------|-----------|---------|--------------|
-| **Ordinary Kriging** | Array of Points `[x, y, value]` and Semivariogram Model | `[predicted value, error of variance, local estimated mean, weights array]` | read data - model semivariogram - predict point at unknown location | Universal method for point interpolation |
-| **Simple Kriging** | Array of Points `[x, y, value]`, process mean $\mu$ and Semivariogram Model | `[predicted value, error of variance, given mean, weights array]` | read data - model semivariogram - predict point at unknown location | Useful for processes where mean is known |
-| **Centroid-based Poisson Kriging** | Polygons with rates, points within areas with rates denominator counts and deconvoluted semivariogram model | `[predicted value, error of variance, given mean, weights array]` | read areal data and point data - perform semivariogram deconvolution - interpolate point (centroid) values | Much faster than Area-to-Area and Area-to-Point Poisson Kriging but introduces bias related to the different shapes and sizes of areas |
-| **Area-to-Area Poisson Kriging** | Polygons with rates, points within areas with rates denominator counts and deconvoluted semivariogram model | `[predicted value for area, error of variance, given mean, weights array]` | read areal data and point data - perform semivariogram deconvolution - interpolate areal rates | Filtering and smoothing choropleth maps |
-| **Area-to-Point Poisson Kriging** | Polygons with rates, points within areas with rates denominator counts and deconvoluted semivariogram model | `[predicted value for point within area, error of variance, given mean, weights array, point coordinates]` | read areal data and point data - perform semivariogram deconvolution - interpolate point within area rates | Change of spatial support of analyzed areal rates |
+| **Ordinary Kriging** | Array of Points *[x, y, value]* and Semivariogram Model | *[predicted value, error of variance, local estimated mean, weights array]* | read data - model semivariogram - predict point at unknown location | Universal method for point interpolation |
+| **Simple Kriging** | Array of Points *[x, y, value]*, process mean $\mu$ and Semivariogram Model | *[predicted value, error of variance, given mean, weights array]* | read data - model semivariogram - predict point at unknown location | Useful for processes where mean is known |
+| **Centroid-based Poisson Kriging** | Polygons with rates, points within areas with rates denominator counts and deconvoluted semivariogram model | *[predicted value, error of variance, given mean, weights array]* | read areal data and point data - perform semivariogram deconvolution - interpolate point (centroid) values | Much faster than Area-to-Area and Area-to-Point Poisson Kriging but introduces bias related to the different shapes and sizes of areas |
+| **Area-to-Area Poisson Kriging** | Polygons with rates, points within areas with rates denominator counts and deconvoluted semivariogram model | *[predicted value for area, error of variance, given mean, weights array]* | read areal data and point data - perform semivariogram deconvolution - interpolate areal rates | Filtering and smoothing choropleth maps |
+| **Area-to-Point Poisson Kriging** | Polygons with rates, points within areas with rates denominator counts and deconvoluted semivariogram model | *[predicted value for point within area, error of variance, given mean, weights array, point coordinates]* | read areal data and point data - perform semivariogram deconvolution - interpolate point within area rates | Change of spatial support of analyzed areal rates |
 
[…]
