+++ paper/paper.md
@@ -20,33 +20,37 @@ bibliography: paper.bib
 
 # Summary
 
-Spatial Interpolation techniques are used to interpolate values at unknown locations and/or filter and smooth existing data sources. Those methods work for point observations and areal aggregates. The basic idea is that measurement at one location is a function of weighted distance between this place and its neighbors. It is known as the Tobler's First Law of Geography, which states: *everything is related to everything else, but near things are more related than distant things* [@Tobler:1970]. Kriging technique designed for mining applications exploits this statement formally and nowadays it has gained a lot of attention outside the initial area of interest. Kriging is rather a set of methods which can be applied to environmental science, hydrogeology, natural resources monitoring, remote sensing, epidemiology and ecology problems [@Chilès:2018]. Kriging is usually used with point measurements or regular block units but many real-world datasets are different. Especially challenging are measurements of rates over areas of irregular shapes and sizes, as example administrative units in every country [@Goovaerts:2007]. Pyinterpolate is designed to tackle this problem and to simplify area variogram regularization and analysis for social, environmental, public health and similar sciences. Package merges basic Kriging techniques with more sophisticated Poisson Kriging methods.
+Spatial Interpolation techniques are used to interpolate values at unknown locations and/or filter and smooth existing data sources. Those methods work for point observations and areal aggregates. The basic idea is that measurement at one location is a function of weighted distance between this place and its neighbors. It is known as the Tobler's First Law of Geography, which states: *everything is related to everything else, but near things are more related than distant things* [@Tobler:1970]. Kriging technique designed for mining applications exploits this statement formally and nowadays it has gained a lot of attention outside the initial area of interest. Today "kriging" is rather a set of methods which can be applied to environmental science, hydrogeology, natural resources monitoring, remote sensing, epidemiology and ecology problems [@Chilès:2018]. Kriging is usually used with point measurements or regular block units but many real-world datasets are different. Especially challenging are measurements of rates over areas of irregular shapes and sizes, as example administrative units in every country [@Goovaerts:2007]. Pyinterpolate is designed to tackle this problem and to simplify area variogram regularization. This property makes Pyinterpolate an interpolation and filtering tool for social, environmental and public health sciences. Package merges basic Kriging techniques with more sophisticated Area-to-Area and Area-to-Point Poisson Kriging methods.
 
 # Statement of need
 
-Pyinterpolate is a Python package for spatial interpolation and it is designed to perform predictions from point measurements and areal aggregates of different sizes and shapes. Pyinterpolate automates tasks performed by spatial statisticians, it helps with data exploration, semivariogram estimation and kriging predictions. Thing which makes Pyinterpolate different from other spatial interpolation packages is an ability to perform Kriging of areas of different shapes and sizes which is extremely important in the context of social, medical and ecological sciences.
+Pyinterpolate is a Python package for spatial interpolation and it is designed to perform predictions from point measurements and areal aggregates of different sizes and shapes. Pyinterpolate automates tasks performed by spatial statisticians, it helps with data exploration, semivariogram estimation and kriging predictions. Thing that makes Pyinterpolate different from other spatial interpolation packages is the ability to perform Kriging of areas of different shapes and sizes and this type of operation is extremely important in the context of social, medical and ecological sciences.
 
 ## Why is areal interpolation important?
 
[…]
