+++ paper/paper.bib
@@ -2,7 +2,9 @@
   author = {Armstrong, Margaret},
   publisher = {Springer},
   title = {Basic Linear Geostatistics},
-  year = {1998}
+  year = {1998},
+  doi={10.1007/978-3-642-58727-6},
+  url={https://link.springer.com/book/10.1007/978-3-642-58727-6}
 }
 
 @article{Goovaerts:2006,
