### 2021-03-28 #2869 @szymon-datalions
Hi @kenohori,

I've taken a look into a data and publications which were used by me as the reference for building this package. And I noticed that area of analysis for breast cancer was much larger than I've initially set it. I used single state but in the reference paper it is Northeastern part of United States created from multiple states. With larger area of interest spatial patterns are more visible so in this version of the paper given example shows stronger spatial correlation and reason why we perform an areal deconvolution is clear.

I've inserted table with each type of Kriging and I've rearranged the paper a little bit. The example data is projected in meters too. Thanks for your comments again, I learn a lot every time when something is improved in the paper :)
