### 2021-03-30 de4ef48
updated error weighting (large update)
