+++ pyinterpolate/semivariance/semivariogram_fit/fit_semivariance.py
@@ -1,8 +1,10 @@
 import csv
 from operator import itemgetter
+import warnings
 
 import numpy as np
 import pandas as pd
+from scipy.optimize import curve_fit
 import matplotlib.pyplot as plt
 
 
@@ -56,6 +58,7 @@ class TheoreticalSemivariogram:
         self.chosen_model_name = None
         self.params = None
         self.model_error = None
+        self.is_weighted = False
 
     # MODELS
 
@@ -211,6 +214,9 @@ class TheoreticalSemivariogram:
             between minimal range of empirical semivariance and maximum range of empirical semivariance.
         """
 
+        if weighted:
+            self.is_weighted = True
+
         # models
         models = {
             'spherical': self.spherical_model,
@@ -236,16 +242,13 @@ class TheoreticalSemivariogram:
         ranges = np.linspace(minrange, maxrange, number_of_ranges)
 
         # Calculate model errors
-        model_errors = [('NULL model', base_error, None)]
+        model_errors = [('Linear (LS) reference model', base_error, None)]
         for model in models:
             optimal_range = self.calculate_range(models[model], ranges, nugget, sill)
 
             # output model
             params = [nugget, sill, optimal_range]
-            if not weighted:
-                model_error = self.calculate_model_error(models[model], params)
-            else:
-                model_error = self.calculate_model_error(models[model], params, True)
+            model_error = self.calculate_model_error(models[model], params)
 
             model_errors.append((model, model_error, params))
             if self.verbose:
@@ -254,7 +257,7 @@ class TheoreticalSemivariogram:
         # Select the best model
         sorted_errors = sorted(model_errors, key=itemgetter(1))
 
-        if sorted_errors[0][0] == 'NULL model':
+        if sorted_errors[0][0] == 'Linear (LS) reference model':
             # This is unlikely case when error estimated as the squared distance between extrapolated values and
             # x axis is smaller than models' errors
 
@@ -264,10 +267,12 @@ class TheoreticalSemivariogram:
             self.theoretical_model = models[model_name]
             self.params = model_params
 
-            print('WARNING: NULL model error is better than estimated models, its value is:', sorted_errors[0][1])
-            print('Chosen model: {}, with value of: {}.'.format(
-                model_name, model_error
-            ))
+            warning_msg = 'WARNING: linear model fitted to the experimental variogram is better than the core models!'
+            warnings.warn(warning_msg)
+            if self.verbose:
+                print('Chosen model: {}, with value of: {}.'.format(
+                    model_name, model_error
+                ))
 
         else:
 
@@ -301,8 +306,8 @@ class TheoreticalSemivariogram:
                                               self.params[2])
         return output_model
 
-
-    def _curve_fit(self, a, b, x):
+    @staticmethod
+    def _curve_fit_function(x, a, b):
         """
         Method fits data into a 1st order polynomial curve where:
             y = a * x + b
@@ -321,32 +326,66 @@ class TheoreticalSemivariogram:
         y = a * x + b
         return y
 
+    def _get_weights(self):
+        """
+        Method creates weights based on the lags for each semivariogram point
+
+        OUTPUT:
+
+        :returns: (numpy array)
+        """
+
+        nh = np.sqrt(self.empirical_semivariance[:, 2])
+        vals = self.empirical_semivariance[:, 1]
+        nh_divided_by_vals = np.divide(nh,
+                                       vals,
+                                       out=np.zeros_like(nh),
+                                       where=vals != 0)
+        return nh_divided_by_vals
+
     def calculate_base_error(self):
         """
-        Method calculates base error as the least squared model of experimental semivariance.
-        """
-        n = len(self.empirical_semivariance[:, 1])
-        zeros = np.zeros(n)
[…]
