### 2021-04-01 f69bdd9
weighting against least squares model - end
