### 2021-04-03 0f3d46f
simple kriging global mean update
