### 2021-04-03 227af8d
tqdm progress bar added for some functions and removed from others
