### 2021-04-04 a05f6dd
ranges parameter set based on the max range and step size
