### 2021-04-16 28285fa
semivariogram reg checks
