### 2021-04-19 #2869 @sdesabbata
Hi @szymon-datalions,

I have started looking into your submission today. I was wondering whether you can clarify something about your example. Is this a bit of analysis that has been done in a published paper as suggested by your comment above?

> I've taken a look into a data and publications which were used by me as the reference for building this package. And I noticed that area of analysis for breast cancer was much larger than I've initially set it. I used single state but in the reference paper it is Northeastern part of United States created from multiple states. With larger area of interest spatial patterns are more visible so in this version of the paper given example shows stronger spatial correlation and reason why we perform an areal deconvolution is clear.

If that's the case, the paper should be referenced in your submission and a comparison of the results should be presented.
