### 2021-04-19 #2869 @szymon-datalions
Hi @sdesabbata ,

One more thing. I've taken a look into [mentioned paper](https://ij-healthgeographics.biomedcentral.com/articles/10.1186/1476-072X-4-31) but methodology and spatial extent is slightly different. Should I modify my example to fit the same path of analysis and then compare results? Now it seems to be pointless because my data is different and extent is larger. This paper was an example for me how big should be an extent.
