### 2021-04-20 #2869 @sdesabbata
Hi @szymon-datalions,

I guess it depends what's the aim of the example. I assume that the aim is to demonstrate that your package does what's supposed to do. I haven't check the software and related tests yet, but I guess it would be good to include an example in the paper that does show that the software produces the expected results. That can be done in two ways: you take an example of analysis conducted with an industry-standard software and show that the output is the same; or you work with a simulated dataset with known characteristics and show that you can recreate a surface from a sample.
