### 2021-05-05 8bfad52
code refactoring in semivariogram calculations
