### 2021-05-05 bc96544
tests passed, weighted semivariance calc refactored
