### 2021-05-10 #2869 @szymon-datalions
> Hi @szymon-datalions,
>
> Thanks for the revised paper! I like the improvements very much.
>
> Focussing on areal interpolation (of which there are fewer open implementations) is a great idea, and the explanation of why this is important helps a lot to clarify the strengths of your software. As before, I think this text is also a very nice introduction to kriging for those who are not familiar with it. Referring to kriging as a set of methods is a nice description.
>
> I don't have much more to add at this point. Just some suggestions related to minor formatting things:
>
> * using \times for matrix sizes
> * broken reference at the end of Spatial Interpolation with Kriging
> * shrinking the table text would make it fit in the text block
> * broken link to the GitHub repository at the end of Dataset

Thanks @kenohori; and I'll correct those formatting issues in the paper :)

@hugoledoux I've performed requested tests and I'll update paper by the end of this week so it'll be (hopefully) all from my side.

Thanks!
