### 2021-05-11 f24b39c
refactored semivariogram plots and initial idw algorithm
