+++ pyinterpolate/transform/tests.py
@@ -1,0 +1,13 @@
+def _does_variogram_exist(theoretical_model):
+    """
+    Function checks if variogram is calculated.
+
+    INPUT:
+
+    :param theoretical_model: (TheoreticalSemivariogram)
+
+    OUTPUT:
+
+    :return: (bool) True if Theoretical Semivariogram has calculated params or False otherwise
+    """
+    pass
