+++ pyinterpolate/semivariance/semivariogram_fit/fit_semivariance.py
@@ -56,7 +56,9 @@ class TheoreticalSemivariogram:
         self.verbose = verbose
         self.theoretical_model = None
         self.chosen_model_name = None
-        self.params = None
+        self.nugget = None
+        self.range = None
+        self.sill = None
         self.model_error = None
         self.is_weighted = False
 
@@ -218,12 +220,14 @@ class TheoreticalSemivariogram:
         optimal_range = self.calculate_range(model, ranges, nugget, sill)
 
         # output model
-        self.params = [nugget, sill, optimal_range]
+        self.nugget = nugget
+        self.sill = sill
+        self.range = optimal_range
 
         # model error
-        self.model_error = self.calculate_model_error(model, self.params)
+        self.model_error = self.calculate_model_error(model, self.nugget, self.sill, self.range)
 
-        return model_type, self.params
+        return model_type
 
     def find_optimal_model(self, weighted=False, number_of_ranges=16):
         """
@@ -277,10 +281,9 @@ class TheoreticalSemivariogram:
             optimal_range = self.calculate_range(models[model], ranges, nugget, sill)
 
             # output model
-            params = [nugget, sill, optimal_range]
-            model_error = self.calculate_model_error(models[model], params)
+            model_error = self.calculate_model_error(models[model], nugget, sill, optimal_range)
 
-            model_errors.append((model, model_error, params))
+            model_errors.append((model, model_error, [nugget, sill, optimal_range]))
             if self.verbose:
                 print('Model: {}, error value: {}'.format(model, model_error))
 
@@ -294,8 +297,6 @@ class TheoreticalSemivariogram:
             model_name = sorted_errors[1][0]
             model_error = sorted_errors[1][1]
             model_params = sorted_errors[1][2]
-            self.theoretical_model = models[model_name]
-            self.params = model_params
 
             warning_msg = 'WARNING: linear model fitted to the experimental variogram is better than the core models!'
             warnings.warn(warning_msg)
@@ -303,18 +304,19 @@ class TheoreticalSemivariogram:
                 print('Chosen model: {}, with value of: {}.'.format(
                     model_name, model_error
                 ))
-
         else:
-
             model_name = sorted_errors[0][0]
             model_error = sorted_errors[0][1]
             model_params = sorted_errors[0][2]
-            self.theoretical_model = models[model_name]
-            self.params = model_params
             if self.verbose:
                 print('Chosen model: {}, with value: {}.'.format(
                     model_name, model_error
                 ))
+
+        self.theoretical_model = models[model_name]
+        self.nugget = model_params[0]
+        self.sill = model_params[1]
+        self.range = model_params[2]
         self.chosen_model_name = model_name
         self.model_error = model_error
         return model_name
@@ -331,9 +333,9 @@ class TheoreticalSemivariogram:
 
     def calculate_values(self):
         output_model = self.theoretical_model(self.empirical_semivariance[:, 0],
-                                              self.params[0],
-                                              self.params[1],
-                                              self.params[2])
+                                              self.nugget,
+                                              self.sill,
+                                              self.range)
         return output_model
 
     @staticmethod
@@ -399,7 +401,7 @@ class TheoreticalSemivariogram:
             mean_error = np.mean(weights * error)
             return mean_error
 
-    def calculate_model_error(self, model, par):
+    def calculate_model_error(self, model, nugget, sill, semivar_range):
         """
         Function calculates error between specific models and experimental curve.
 
@@ -408,9 +410,9 @@ class TheoreticalSemivariogram:
         :returns: (float) mean squared difference between model and experimental variogram.
         """
         error = np.sqrt((self.empirical_semivariance[:, 1] - model(self.empirical_semivariance[:, 0],
-                                                                   par[0],
-                                                                   par[1],
-                                                                   par[2])) ** 2)
[…]
