### 2021-05-13 cece8ce
check variogram method
