+++ pyinterpolate/kriging/point_kriging/kriging.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from pyinterpolate.distance.calculate_distances import calc_point_to_point_distance
 from pyinterpolate.transform.prepare_kriging_data import prepare_kriging_data
-from pyinterpolate.transform.tests import _does_variogram_exist
+from pyinterpolate.transform.tests import does_variogram_exist
 
 
 class Krige:
@@ -17,10 +17,10 @@ class Krige:
     - ordinary_kriging - ordinary kriging of unknown point value,
     - simple_kriging - simple kriging of unknown point value.
 
-    Class methods may raise ValueError if estimated value is below 0. You may use try: ... except: ... statement to overwrite those values
-        with some constant or NaN or you could use different semivariogram model. Sometimes this problem is related to
-        the input data, especially to the clustered groups of points. In this case aggregate those clusters and then estimate
-        semivariogram and perform kriging.
+    Class methods may raise ValueError if estimated value is below 0. You may use try: ... except: ... statement to
+        overwrite those values with some constant or NaN or you could use different semivariogram model. Sometimes this
+        problem is related to the input data, especially to the clustered groups of points. In this case aggregate
+        those clusters and then estimate semivariogram and perform kriging.
 
     INITLIALIZATION PARAMS:
 
@@ -34,11 +34,15 @@ class Krige:
         INPUT:
 
         :param semivariogram_model: (TheoreticalSemivariogram) Theoretical Semivariogram used for data interpolation,
-        :param known_points: (numpy array) dataset with known values and locations. Each column should represent different dimension and the last column represents values
-        example: [[dim_x1, dim_y1, val_1], [dim_x2, dim_y2, val_2]]
+        :param known_points: (numpy array) dataset with known values and locations. Each column should represent
+            different dimension and the last column represents values:
+            [[dim_x1, dim_y1, val_1], [dim_x2, dim_y2, val_2]]
         """
 
         self.dataset = known_points
+
+        # Test semivariogram model
+        does_variogram_exist(semivariogram_model)
         self.model = semivariogram_model
         self.distances = None
 
