### 2021-05-13 7cb746f
fit_semivariance tests fixes
