+++ docs/build/html/_sources/code_documentation/semivariance.ipynb.txt
@@ -782,7 +782,7 @@
     "TheoreticalSemivariogram.import_model(self, filename)\n",
     "```\n",
     "\n",
-    "Function imports semivariance model and updates it's parameters (model name, nugget, sill, range, model_error).\n",
+    "Function imports semivariance model and updates its parameters (model name, nugget, sill, range, model_error).\n",
     "\n",
     "---\n",
     "\n",
@@ -813,13 +813,6 @@
     "```\n",
     "\n",
     "Function shows experimental and theoretical semivariogram in one plot."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "---"
    ]
   }
  ],
