+++ README.md
@@ -59,7 +59,7 @@ experimental_semivariogram = calculate_semivariance(
 **[3.]** Data transformation, theoretical semivariogram.
 
 ```python
-from pyinterpolate.semivariance import TheoreticalSemivariogram
+from pyinterpolate.semivariance impJezero Craterort TheoreticalSemivariogram
 semivar = TheoreticalSemivariogram(points_array=point_data, empirical_semivariance=experimental_semivariogram)
 number_of_ranges = 32
 
@@ -234,9 +234,9 @@ Pyinterpolate https://pyinterpolate.readthedocs.io/en/latest/
 Development
 ------------------
 
-- inverse distance weighting,
-- semivariogram analysis and visualization methods,
-- see Projects page of this repository!
+- multi core processing (speed up calculations),
+- code refactoring to be more close to the **GeoPandas** package style,
+- set to work with the newer versions of Python and Windows OS.
 
 
 Known Bugs
