+++ paper/paper.bib
@@ -131,6 +131,34 @@ url="https://doi.org/10.1007/978-3-319-78999-6_29"
   url          = {https://doi.org/10.5281/zenodo.4385980}
 }
 
+@article{PebesmaSPArt,
+    author = {Edzer J. Pebesma and Roger S. Bivand},
+    title = {Classes and methods for spatial data in {R}},
+    journal = {R News},
+    year = {2005},
+    volume = {5},
+    number = {2},
+    pages = {9--13},
+    month = {November},
+    url = {https://CRAN.R-project.org/doc/Rnews/},
+  }
+
+@book{BivandBook,
+    author = {Roger S. Bivand and Edzer Pebesma and Virgilio
+      Gomez-Rubio},
+    title = {Applied spatial data analysis with {R}, Second edition},
+    year = {2013},
+    publisher = {Springer, NY},
+    url = {https://asdar-book.org/},
+  }
+
+@article{meuse,
+    author = {Edzer J. Pebesma},
+    title = {The meuse data set: a tutorial for the gstat R package},
+    year = {2021},
+    url = {https://cran.r-project.org/web/packages/gstat/vignettes/gstat.pdf},
+}
+
 @article{PEBESMA2004683,
 title = "Multivariable geostatistics in S: the gstat package",
 journal = "Computers & Geosciences",
