### 2021-05-15 b450745
paper table update small font size
