### 2021-05-15 #2869 @szymon-datalions
Hi,

I've made a major update in the paper and the package.

- @sdesabbata I've included comparison to `gstat` Kriging interpolation at the end of a paper (new chapter before Bibliography section),
- @kenohori Thank you for your help, I've updated paper accordingly, but there's a problem with a table width. Locally I'm able to control it with HTML tags but after compilation table no longer fits into the page. @hugoledoux are there any guidelines how to format tables in a paper?

- I've made few changes in figures, to save some space,

This is all from my side. I'm waiting for further comments or acceptance of the paper,

Thanks a lot,
Szymon
