### 2021-05-17 #2869 @hugoledoux
@szymon-datalions while being precise in the paper is important, now the paper is rather long, way longer than typical papers. Could we make sure that it stays <10pages? You could for instance merge figures 2-3 and 4-5 together, and maybe the use-cases could be in a wiki in the repository or in the docs? The point is to "Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.", but all the details should be moved somewhere else.

The point of JOSS is having a *short* paper, which is just a windows on the code/docs. See what it should contain: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
