### 2021-05-17 #2869 @hugoledoux
>     * @hugoledoux are there any guidelines how to format tables in a paper?

hmmm, I am not sure how. You can resize figures like this, perhaps this works for tables?
```
![figure capction](extrusion.png){ width=90% }
```
If not we'll deal with this at the end.
