### 2021-05-17 #2869 @szymon-datalions
@hugoledoux I've changed paper - I moved two big blocks into separate files in the paper repository. One block is a use case and second block is a comparison of methods between two different packages but focused mostly on the specific solution. The main parts of paper are still here and I've managed to shrink it to 9. pages.

Table is still larger than the text field, this method don't work. I'll try to find a solution - maybe I'll insert table as an image to avoid problems with formatting. But maybe this part should be done later? I still need one more review to go.
