### 2021-06-12 #2869 @szymon-datalions
@hugoledoux I've removed table from the paper and create numbered list from it. I don't know why but in the previous compilations `wheadon` has compiled paper from the custom branch `paper` but after while it was changed to the version from the `main` branch. You may check timestamps: I've updated paper in the custom branch `paper` and three minutes later I've initialized `wheadon`. When I had checked paper everything was ok but after few minutes I checked it again and I saw version from the `main` branch.

So I updated `main` branch to overcome this problem but I don't know if this is something on my side or something with `wheadon`.

@sdesabbata now paper is really short and some parts are moved to the supplementary materials (comparison of capabilities and output, use case). If you want to check comparison of the package capabilities to existing software it is described here: https://github.com/szymon-datalions/pyinterpolate/tree/main/paper/supplementary%20materials
