### 2021-06-28 #2869 @szymon-datalions
@sdesabbata Hi, I've restructured paper slightly to make it more clear and accessible. It is not a big change, rather I gave it the finishing touch.
