### 2021-07-09 #2869 @sdesabbata
As agreed with @hugoledoux I am waiting for a colleague of my with regard to the discussion of Kriging. In the meantime, here are some detailed feedback on the current version of the paper.

The key point that requires further clarification in my view is the automation of the analysis, which is mainly discussed in line 98 to 105. The outcome of a Kriging process are based on a number of assuptions made by the spatial analyst about the underlying process. Changing assumptions can significantly change the output. In particular, the semivariogram model (I assume that is what you refer to by when you say "teoretical model") is crucial. The paper should provide further details about how such process is automated. That is mentioned briefly in the paragraph starting in line 130 and in the case studies. A more clear reference from the paragraph starting in line 98 to the one starting in line 130 should be made, and further details should be provided. I guess the automation is particularly concerning when drift is present in the data, and an estimation of the errors might not be as good the analysis of an analyst. That might also relate to the process of outliers removal (line 171), where the paper would also benefit from some further details.

With regard to the general structure, would it not make sense to discuss Kriging as a method first (section "Method") and then the implementation (section "69 Interpolation methods within Pyinterpolate") before the details of the modules (section "Modules")?

Try be consistent with regard to the use of the terminology. Also, try to separate the discussion about data and the transformation of the data from the use of the term "map", where the latter is a representation of the data, not the data themselves.

- README.md, Recommended - conda installation, step 2, has the syntax not been updated to eliminate "env"?
- Line 31, seems to suggest that interpolation is done manually. Please rephrase.
- Line 38 (and later in the paper), I am not familiar with the term "point support map", do you mean maps based on point-support model?
- Line 41, I understand what you mean by transforming "choropleth maps", but I think in this case that's not appropriate as it is the area unit data which are being transformed, not the map, which is just a representation.
- Line 43, point 1 needs to be explained better.
- Line 47, point 3 needs to be explained better.
- Line 53, "chained-models systems" requires a reference.
- Line 55, it is unclear what the purpose of the procedure there described would be. Is that meant as an alternative to area-to-area Kriging?
- Line 63, this is a good example, but it would be good to expand the descirption a bit more and clearify it.
- Line 73, simple Kriging also assumes that the mean doesn't vary spatially in a systematic way.
- Line 86, how does the algorithm choses the theoretical model? Which algorithm is used to make the selection?
- Line 91, what do you mean by "counts over area"?
- Line 92, what do you mean by "on the other spectrum"? What do you mean by "universal technique"? The latter is particularly confusing in this context, as one of the Kriging approaches is "universal Kriging".
- Line 93, what do you mean by "multiple point data sources"?
- Line 98, which steps? The rest of the paragraphs says something about which steps can be controlled by the user, but which one are completely out of the user's control?
- Line 196, what do you mean by "areal scale"?
- Figure 1, the use a diverging red-to-blue colour scheme for the first map is confusing. Please use a single-hue scheme as done in the other maps to make the map easier to compare with the others.
- General notes on grammar.
  - the use of articles needs to be revised. The article "the" in particular is frequently missing (e.g., line 32 should be "the thing that makes...", or it is used instead of "a" (e.g. line 29 should be "as a noise-filtering algorithm".
