### 2021-07-10 #2869 @szymon-datalions
@sdesabbata Thanks for your feedback. I'll go point by point and I'll let you know about changes / additional explanations / corrections when they'll be done.
