### 2021-07-16 #2869 @hugoledoux
One point I'd like to make clear: the paper should in general be kept short, now it's 9p with the references and it shouldn't be made longer, ideally.

The recommended errors/typos and clarifications proposed by @sdesabbata should be implemented, but adding new sections about how certain parts (eg "The paper should provide further details about how such process is automated.") could be put in the docs of the software package. The paper is supposed to introduce the package and explain why it is necessary, JOSS prefers that the details of the package and how to use it are in the docs.
