### 2021-08-24 #2869 @szymon-datalions
@sdesabbata I've updated paper with regard to your comments and advices. It's slightly longer, but I've created two additional files within the package documentation related to the Outliers Detection and to the Semivariance fitting process. Links to those files are given in the appendix. I've sent some time with the grammar too and I hope that now it's much better than before.

Thank you
