### 2021-11-05 #2869 @hugoledoux
@kenohori & @sdesabbata could you resume the review and update us? The author updated the paper and the package.

Please let us know
