### 2021-11-06 #2869 @kenohori
I was already happy with the previous version of the paper, so I only had a quick look at the updated version. It is fine by me: reads well and has all required elements. I assume this is sufficient? Or do you want me to test the updated software?
