### 2021-11-19 #2869 @sdesabbata
Thank you, @SimonMolinsky, for the edits to the paper. I think the new version reads much better, and it provides an improved showcase of the package. However, in my second reading, I have found two issues, one of which took me some time to clarify.

The first issue relates to the *Spatial Interpolation with Kriging* section. There seems to be an error in equation 2 and 3, or rather in their description. Line 90 says that `mu` is *"a process mean"* (which is in itself unclear as a sentence), but I believe that is a Lagrange multiplier. Armstrong (1998) uses the same notation as equation 3 in her equation 7.9, and she describes it as the Lagrange multiplier. Her equation 7.11 is a bit different from equation 2, but equation 2 seems correct based on other sources, assuming that `mu` is the Lagrange multiplier rather than *"a process mean"*.

I have looked at the code for the related functions in the package. As far as I can tell, the code seems fine, but I think it is fundamental that you double-check, and I would be grateful if the other reviewers could do the same.

There are some other minor issues with notation. Most summations don't include an upper bound, and `gamma` should be included as the left part of equation 4 to clarify it.

The second issue relates to the *Ordinary Kriging of meuse dataset* example in the additional materials. The example is a very good showcase of the package, and the `R` `gstat` package is a great choice. However, the scatterplot and final maps illustrate patterns (highlighted in red in the images below) in the `pyinterpolate` output that I believe would require further investigation. Furthermore, the two error maps seem to produce a similar spatial pattern but with different distributions. However, that is difficult to assess without a legend. I suggest adding a legend for each one of the four maps in the notebook and further investigating the difference and patterns in the errors.
