### 2021-11-22 #2869 @hugoledoux
> I have looked at the code for the related functions in the package. As far as I can tell, the code seems fine, but I think it is fundamental that you double-check, and I would be grateful if the other reviewers could do the same.

@sdesabbata could you please tick the boxes above so that we can clearly identify what is left?

And the code has been checked already by the other reviewer, so I assume this is okay.

- - -

@SimonMolinsky can you address the notation issue raised?  And I agree that adding legends to the maps is required.
