### 2021-11-22 #2869 @SimonMolinsky
@sdesabbata,

Thank you for your comments. You were right with the description of Lagrange Multiplier, it was a mistake that I've described this parameter as a "process mean". I've checked code to be sure that the implementation doesn't use any "mean" too :) and there is a correct Lagrange Multiplier in the final equation of Ordinary Kriging.

I'll change descriptions and notations, add legends to the maps and validate those patterns in a matter of few days.
