### 2021-11-22 #2869 @sdesabbata
Thank you @SimonMolinsky. I have re-run the example and added the legends to better review the results. I have created a pull request for the paper repo.

@hugoledoux Based on the replies above, the presence of those patters is the only concern I have at the moment.
