### 2021-11-29 #2869 @SimonMolinsky
Hi @sdesabbata ,

I've found the root cause of the correlogram patterns. It is related to the theoretical variogram parameters (sill and range). If I use parameters from the `gstat` package, I obtain a much closer correlogram. I've updated the notebook with it; you can compare both theoretical models: https://github.com/SimonMolinsky/pyinterpolate-paper/blob/main/paper-examples/example-gstat-comparison/Ordinary%20Kriging%20of%20Meuse%20dataset.ipynb

I think that the rest of the differences are a product of the nearest neighbors selection and a different lag calculation. The range of lag in `gstat` is an interval between two distances (lags). Ranges differ in size. In `pyinterpolate` range is estimated from lag and halved distance to the previous lag and a halved distance to the next lag (so lag distance is a point in the middle). Ranges are equal.

I don't know if I should force the package to work as `gstat` - I will ensure that the user may provide their list of lags (it seems to be more important at this stage).

I'm more concerned about the error output (patterns in error maps). I must investigate it further, especially why changes in patterns are so abrupt. To do so, I must take a look into `gstat` implementation of Kriging. Again, I think that this issue may be related to the neighbors selection in `pyinterpolate`.
