### 2021-12-06 #2869 @SimonMolinsky
Hi @sdesabbata ,

I've found why the error map is wrong - and it affects much more than errors only. It is related to the function that picks the closest neighbors:

```python
def prepare_kriging_data(unknown_position, data_array, number_of_neighbours=10):
    """
    Function prepares data for kriging - array of point position, value and distance to an unknown point.
    INPUT:
    :param unknown_position: (numpy array) position of unknown value,
    :param data_array: (numpy array) known positions and their values,
    :param number_of_neighbours: (int) number of the closest locations to the unknown position.
    OUTPUT:
    :return: (numpy array) dataset with position, value and distance to the unknown point:
        [[x, y, value, distance to unknown position], [...]]
    """

    # Distances to unknown point
    r = np.array([unknown_position])

    known_pos = data_array[:, :-1]
    dists = calc_point_to_point_distance(r, known_pos)

    # Prepare data for kriging
    kriging_output_array = np.c_[data_array, dists.T]
    kriging_output_array = kriging_output_array[kriging_output_array[:, -1].argsort()]
    prepared_data = kriging_output_array[:number_of_neighbours]

    return prepared_data
```

This function takes n-closest neighbors instead of **the closest** neighbors based on the spatial distance. I've performed a few experiments and obtained results similar to `gstat`. I'll update the package and paper example accordingly and let you know when it's done. Thanks!
