### 2021-12-15 #2869 @SimonMolinsky
@sdesabbata ,

problem with the strange error map results was more profound than the range selection. It was related to the code logic: error variance was calculated only based on the closest point weight... For most cases, this weight was enough to get comparable results, but for the points' clusters and multiple points within the very close distance, it was giving wrong results... And this part of the package was not extensively tested by me; I assumed that if predictions are good enough, then error variance must be good too...

Here is the new comparison:

![errors_comparison](https://user-images.githubusercontent.com/31246246/146179182-c5f9ce25-4894-41b3-8749-528e3c695056.png)

And the absolute difference:

![errors_difference](https://user-images.githubusercontent.com/31246246/146179243-8cb2fe69-4edb-4199-b52d-2543916542a2.png)

As you may see now results are very close. Minor differences are related to the different range division: **pyinterpolate** divides the area into equally-spaced lags and **gstat** divides the area based on the optimization of the variogram model (so lags are close but not equal).

I've updated the package. Many thanks for pointing out those strange results within the error maps!

The Paper is updated too. Equations and their description is now correct. I'm looking for the revision update. Thanks a lot!
