### 2022-02-01 #2869 @hugoledoux
@SimonMolinsky I am sorry that it took so long. Since you modified the code according to the reviewers wishes, I think we can move forward with the acceptance of the paper.

I just went over it and created a PR to fix some issues: https://github.com/DataverseLabs/pyinterpolate/pull/203

Also, I think it it still too long. As written above, JOSS favours short papers and details should be put in the docs. So I propose the following:

  1. Remove the Kriging equations from the paper, these should be put somewhere in the docs. You can refer to this in the paper.
  2. I think that the sections "Areal data transformation" (l.165-203) is not necessary in the paper. Again, should be added somewhere in the docs.
  3. I would leave Fig2, but remove the text afterwards and just leave the packages' names speak for themselves. So I'd delete l.208-229.
  4. Appendix is not necessary, the paper format already has links to the code (I deleted those in my PR already).

Let me know when you have done this changes.
