### 2022-02-01 #2869 @SimonMolinsky
@chrisbrunsdon thank you for your support!

I've made all changes:

1. Has been moved into supplementary md file within repo, link given.
2. As above.
3. Done.
4. Done.
5. Ok.

I've made few cosmetic changes in the paper:

- Lines 89-91: Description of ATA PK, it was wrong,
- Line 97: Resources with the theory of Kriging
- Lines 131-137: Grouped external resources with the additional context.

I hope that everything is ok by now.
