### 2022-02-20 #2869 @SimonMolinsky
Hi @hugoledoux ,

I don't know if my reply on the same day as the change request was overlooked, but maybe I will do it again :) I've made requested changes in a paper. It has ~4 pages of text if we exclude one large figure and bibliography. I hope that it's ok and paper fits JOSS with this structure.
