### 2022-02-21 #2869 @hugoledoux
Sorry @SimonMolinsky indeed you replied too fast and I didn't notice it and then I was on holiday.

I like the update you made, just went over it. I have a few small final fixes: https://github.com/DataverseLabs/pyinterpolate/pull/209

When this is done, I am happy to start the final steps for the formal acceptance.

At this point could you:
