### 2022-02-21 #2869 @SimonMolinsky
Hi @hugoledoux

Here are:

- [x] Version tag: [0.2.5.post1](https://github.com/DataverseLabs/pyinterpolate/releases/tag/v0.2.5.post1)
- [x] Zenodo Archive: [10.5281/zenodo.6206145](https://zenodo.org/record/6206145#.YhOxci-pn5k)

I don't understand why did I send package to Zenodo? Should I update it later with the new versions?

Thank you.
