### 2022-02-22 #2869 @hugoledoux
You got a DOI because your paper is accepted and we're finalising everything!

One small typo to fix: in the zenodo the capital "i" of "Interpolation" is still there, and it shouldn't... we need to have the exact same title. Please fix this, then I can process with the formal acceptance (you might be asked to fix things I missed, the editor-in-chief does that part)
