### 2022-02-22 #2869 @SimonMolinsky
I've [fixed this typo](https://zenodo.org/record/6206145#.YhTLBi-pm_w), we can move forward :) Thanks!
