### 2022-02-22 #2869 @danielskatz
@hugoledoux - just a small correction:

"JOSS archives your paper, but not your software; we need you to do this yourself, but then we link to it, and this is one of the last steps before your paper is recommended for acceptance"
