### 2022-02-23 #2869 @SimonMolinsky
The great news for today!

Thank you @hugoledoux for all your help and supervision along the way, and your work for the Open Source community,

Thank you @kenohori and @sdesabbata for your time, patience, and invaluable comments, you've helped me a lot! Without your insights the package will be unreliable.

Thank you @arfon for acceptance and publication.

I'm sure that my journey with JOSS just stared ;) and it's just the first step. I'd gladly join to the reviewers team and donate for JOSS.

Cheers! 🍾 🥳
